Move base app to calchart app
