--- calchart/base/__init__.py
@@ -1,1 +1,0 @@
-"""The package for the base app."""
