+++ calchart/calchart/templatetags/partials.py
