+++ calchart/settings/base.py
@@ -11,7 +11,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.staticfiles',
     'storages',
-    'base',
+    'calchart',
     # 'wiki',
 ]
 
@@ -58,7 +58,7 @@ STATICFILES_DIRS = (
     os.path.join(BASE_DIR, 'static'),
 )
 
-AUTH_USER_MODEL = 'base.User'
+AUTH_USER_MODEL = 'calchart.User'
 
 LOGIN_REDIRECT_URL = 'home'
 LOGIN_URL = 'login'
